refactor auth types
