+++ frontend/src/components/account-settings.ts
@@ -3,9 +3,10 @@ import { state, query, property } from "lit/decorators.js";
 import { msg, localized } from "@lit/localize";
 import { createMachine, interpret, assign } from "@xstate/fsm";
 
-import type { AuthState, CurrentUser } from "../types/auth";
+import type { CurrentUser } from "../types/auth";
 import LiteElement, { html } from "../utils/LiteElement";
 import { needLogin } from "../utils/auth";
+import type { AuthState } from "../utils/AuthService";
 import AuthService from "../utils/AuthService";
 
 @localized()
