+++ frontend/src/pages/join.ts
@@ -2,7 +2,7 @@ import { state, property } from "lit/decorators.js";
 import { msg, localized, str } from "@lit/localize";
 import { createMachine, interpret, assign } from "@xstate/fsm";
 
-import type { AuthState } from "../types/auth";
+import type { AuthState } from "../utils/AuthService";
 import LiteElement, { html } from "../utils/LiteElement";
 import type { LoggedInEvent } from "../utils/AuthService";
 import AuthService from "../utils/AuthService";
