+++ frontend/src/types/auth.ts
@@ -1,12 +1,3 @@
-export type Auth = {
-  username: string;
-  headers: {
-    Authorization: string;
-  };
-};
-
-export type AuthState = Auth | null;
-
 export type CurrentUser = {
   id: string;
   email: string;
